refactor: move d3fc-element examples
