+++ examples/element-dual-axes/index.html
@@ -1,0 +1,39 @@
+<html>
+
+<head>
+    <script src="../../node_modules/seedrandom/seedrandom.js"></script>
+    <script>Math.seedrandom('a22ebc7c488a3a47');</script>
+    <script src="../../node_modules/d3/dist/d3.js"></script>
+    <script src="../../packages/d3fc/build/d3fc.js"></script>
+    <style>
+        body {
+            display: flex;
+        }
+
+        body>* {
+            flex: auto;
+        }
+    </style>
+</head>
+
+<body>
+    <d3fc-group id="chart" auto-resize style="display: flex; height: 40vw; width: 60vw; flex-direction: column">
+        <h1 style="text-align: center">
+            Dual Axes Chart Layout
+        </h1>
+        <div style="height: 1em; display: flex; flex-direction: row">
+            <d3fc-svg id="x-axis-top" style="flex: 1; margin: 0 3em"></d3fc-svg>
+        </div>
+        <div style="flex: 1; display: flex; flex-direction: row">
+            <d3fc-svg id="y-axis-left" style="width: 3em"></d3fc-svg>
+            <d3fc-svg id="plot-area" style="flex: 1"></d3fc-svg>
+            <d3fc-svg id="y-axis-right" style="width: 3em"></d3fc-svg>
+        </div>
+        <div style="height: 1em; display: flex; flex-direction: row">
+            <d3fc-svg id="x-axis-bottom" style="flex: 1; margin: 0 3em"></d3fc-svg>
+        </div>
+    </d3fc-group>
+    <script src="index.js"></script>
+</body>
+
+</html>
