--- packages/d3fc-element/examples/dual-axes.html
@@ -1,30 +1,0 @@
-<!doctype html>
-<html>
-<body>
-
-  <d3fc-group id="chart" auto-resize style="display: flex; height: 40vw; width: 60vw; flex-direction: column">
-    <h1 style="text-align: center">
-      Dual Axes Chart Layout
-    </h1>
-    <div style="height: 1em; display: flex; flex-direction: row">
-      <d3fc-svg id="x-axis-top" style="flex: 1; margin: 0 3em"></d3fc-svg>
-    </div>
-    <div style="flex: 1; display: flex; flex-direction: row">
-      <d3fc-svg id="y-axis-left" style="width: 3em"></d3fc-svg>
-      <d3fc-svg id="plot-area" style="flex: 1"></d3fc-svg>
-      <d3fc-svg id="y-axis-right" style="width: 3em"></d3fc-svg>
-    </div>
-    <div style="height: 1em; display: flex; flex-direction: row">
-      <d3fc-svg id="x-axis-bottom" style="flex: 1; margin: 0 3em"></d3fc-svg>
-    </div>
-  </d3fc-group>
-
-  <script src="../node_modules/babel-polyfill/dist/polyfill.js"></script>
-  <script src="../node_modules/custom-event-polyfill/custom-event-polyfill.js"></script>
-  <script src="../node_modules/document-register-element/build/document-register-element.js"></script>
-  <script src="../node_modules/d3/dist/d3.js"></script>
-  <script src="../../d3fc/build/d3fc.js"></script>
-  <script src="dual-axes.js"></script>
-
-</body>
-</html>
